pyproject.toml
[tool.poetry]
name = "ghmeister"
version = "0.1.0"
description = "GitHub Manager"
authors = ["Rob van der Leek <5324924+robvanderleek@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
ghm = "ghmeister.__main__:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
typer = "^0.9.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
rich = "^13.7.1"
sh = "^2.0.6"
inquirerpy = "^0.3.4"
pydantic = "^2.8.2"

[tool.poe.tasks.bundle]
help = "Create a binary executable using pyinstaller"
cmd = "pyinstaller --workpath .build --noconfirm --specpath dist -n ghmeister ghmeister/__main__.py"

[tool.semantic_release]
branch = "main"
version_toml = [ "pyproject.toml:tool.poetry.version" ]
version_variables = [ "ghmeister/version.py:version" ]
upload_to_repository = false
upload_to_release = false

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.2.0"
poethepoet = "^0.24.4"
pytest = "^7.2.1"
mkdocs-material = "^9.5.10"
termynal = "^0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
